+++ third-party/wocky-client/src/transport/Transport.ts
@@ -1047,6 +1047,43 @@ export class Transport {
     })
   }
 
+  async userBulkLookup(phoneNumbers: string[]): Promise<any[]> {
+    const res = await this.client!.query<any>({
+      query: gql`
+          query userBulkLookup($phoneNumbers: [String]!) {
+            userBulkLookup(phoneNumbers: $phoneNumbers) {
+              e164PhoneNumber
+              phoneNumber
+              user {
+                ${PROFILE_PROPS}
+              }
+              relationship
+            }
+          }
+        `,
+      variables: {phoneNumbers},
+    })
+    const results = res.data.userBulkLookup
+    return results.map(r => ({...r, user: r.user ? convertProfile(r.user) : null}))
+  }
+
+  async friendSmsInvite(phoneNumber: string): Promise<void> {
+    return this.voidMutation({
+      mutation: gql`
+        mutation friendBulkInvite(
+          $phoneNumbers: [String!]
+        ) {
+          friendBulkInvite(
+            input: {phoneNumbers: $phoneNumbers}
+          ) {
+            ${VOID_PROPS}
+          }
+        }
+      `,
+      variables: {phoneNumbers: [phoneNumber]},
+    })
+  }
+
   /******************************** SUBSCRIPTIONS ********************************/
 
   subscribeSharedLocations() {